MGNLUI-3542 Null handling in BasicUploadFieldFactory class, modify version in update task, modify tests.

+++ magnolia-ui-form/src/main/java/info/magnolia/ui/form/field/factory/BasicUploadFieldFactory.java
@@ -80,7 +80,9 @@ public class BasicUploadFieldFactory extends AbstractFieldFactory<BasicUploadFie
         this.admincentralEventBus.addHandler(AppLifecycleEvent.class, new AppLifecycleEventHandler.Adapter() {
             @Override
             public void onAppStopped(AppLifecycleEvent event) {
-                field.getValue().getFile().delete();
+                if (field.getValue().getFile() != null) {
+                    field.getValue().getFile().delete();
+                }
             }
         });
         return uploadField;
